requirements.txt
matplotlib
numpy
torch
torchvision
scikit-learn
seaborn
tqdm
grad-cam
opencv-python-headless
gradio
Pillow